feat: remove the need to make ActionState's copy-able after verification.
This is done by a two-step `onAction` trigger, the first will register the verification checks and re-trigger `onAction`. This will verify and trigger the action if valid. Regardless, the checks will be removed after the attempt, and the `onAction` reset to the initial condition. Each attempt to call onAction will result in a new state instance